# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "OCPP Central System API",
        "description": "REST API for managing OCPP 1.6 charge points and transactions",
        "contact": {
            "name": "Texnouz",
            "email": "support@texnouz.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/charge-points/{charge_point_id}": {
            "get": {
                "tags": [
                    "Charge Points"
                ],
                "summary": "Get charge point by ID",
                "operationId": "get_charge_point",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Charge point details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Charge Points"
                ],
                "summary": "Delete charge point",
                "operationId": "delete_charge_point",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Charge point deleted"
                    },
                    "404": {
                        "description": "Charge point not found"
                    }
                }
            }
        },
        "/api/v1/charge-points/{charge_point_id}/change-availability": {
            "post": {
                "tags": [
                    "Commands"
                ],
                "summary": "Change connector availability",
                "operationId": "change_avail",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeAvailabilityRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Command result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point offline or not found"
                    }
                }
            }
        },
        "/api/v1/charge-points/{charge_point_id}/configuration": {
            "get": {
                "tags": [
                    "Commands"
                ],
                "summary": "Get configuration from charge point",
                "operationId": "get_config",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keys",
                        "in": "query",
                        "description": "Comma-separated list of configuration keys",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration values",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point offline or not found"
                    }
                }
            }
        },
        "/api/v1/charge-points/{charge_point_id}/remote-start": {
            "post": {
                "tags": [
                    "Commands"
                ],
                "summary": "Remote start transaction",
                "operationId": "remote_start",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoteStartRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Command result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point offline or not found"
                    }
                }
            }
        },
        "/api/v1/charge-points/{charge_point_id}/remote-stop": {
            "post": {
                "tags": [
                    "Commands"
                ],
                "summary": "Remote stop transaction",
                "operationId": "remote_stop",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoteStopRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Command result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point offline or not found"
                    }
                }
            }
        },
        "/api/v1/charge-points/{charge_point_id}/reset": {
            "post": {
                "tags": [
                    "Commands"
                ],
                "summary": "Reset charge point",
                "operationId": "reset_charge_point",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Command result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point offline or not found"
                    }
                }
            }
        },
...
        "/api/v1/charge-points/{charge_point_id}/trigger-message": {
            "post": {
                "tags": [
                    "Commands"
                ],
                "summary": "Trigger message from charge point",
                "operationId": "trigger_msg",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TriggerMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Command result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point offline or not found"
                    }
                }
            }
        },
        "/api/v1/charge-points/{charge_point_id}/unlock-connector": {
            "post": {
                "tags": [
                    "Commands"
                ],
                "summary": "Unlock connector",
                "operationId": "unlock",
                "parameters": [
                    {
                        "name": "charge_point_id",
                        "in": "path",
                        "description": "Charge point ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UnlockConnectorRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Command result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge point offline or not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/T"
                            }
                        ],
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
            "ChangeAvailabilityRequest": {
                "type": "object",
                "description": "Change availability request",
                "required": [
                    "connector_id",
                    "type"
                ],
                "properties": {
                    "connector_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Connector ID (0 = entire charge point)",
                        "minimum": 0
                    },
                    "type": {
                        "type": "string",
                        "description": "Availability type: \"Operative\" or \"Inoperative\""
                    }
                },
                "example": {
                    "connector_id": 0,
                    "type": "Operative"
                }
            },
...
            "RemoteStartRequest": {
                "type": "object",
                "description": "Remote start transaction request",
                "required": [
                    "id_tag"
                ],
                "properties": {
                    "connector_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Optional connector ID (1-based)",
                        "nullable": true,
                        "minimum": 0
                    },
                    "id_tag": {
                        "type": "string",
                        "description": "RFID tag or identifier for authorization"
                    }
                },
                "example": {
                    "connector_id": 1,
                    "id_tag": "RFID001"
                }
            },
            "RemoteStopRequest": {
                "type": "object",
                "description": "Remote stop transaction request",
                "required": [
                    "transaction_id"
                ],
                "properties": {
                    "transaction_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Transaction ID to stop"
                    }
                },
                "example": {
                    "transaction_id": 123
                }
            },
            "ResetRequest": {
                "type": "object",
                "description": "Reset request",
                "required": [
                    "type"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "description": "Reset type: \"Soft\" or \"Hard\""
                    }
                },
                "example": {
                    "type": "Soft"
                }
            },
...
            "TriggerMessageRequest": {
                "type": "object",
                "description": "Trigger message request",
                "required": [
                    "message"
                ],
                "properties": {
                    "connector_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Optional connector ID",
                        "nullable": true,
                        "minimum": 0
                    },
                    "message": {
                        "type": "string",
                        "description": "Message type: \"BootNotification\", \"Heartbeat\", \"StatusNotification\", etc."
                    }
                },
                "example": {
                    "connector_id": 1,
                    "message": "StatusNotification"
                }
            },
            "UnlockConnectorRequest": {
                "type": "object",
                "description": "Unlock connector request",
                "required": [
                    "connector_id"
                ],
                "properties": {
                    "connector_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Connector ID to unlock (1-based)",
                        "minimum": 0
                    }
                },
                "example": {
                    "connector_id": 1
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key"
            },
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer token"
            }
        }
    },
...
}